Rename shared to common
